fixes integration test whereby injected service was null.
The fix here is that the default rowfactory needs to inject services into the newly created view model.  Previously this was done by way of using the different API (DomainObjectContainer#newViewModel or something like).  Now, though, we simply instantiate, so we need to inject the services explicitly.

The slight complication here is that the servicesInjector service itself needs to be given to the default rowfactory.  We do this by having it implement
ServicesInjectorAware and then have the excel service push the services injector into the row factory if required before it is first used